pyproject.toml
[project]
name = "talent-platform-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "colorlog>=6.9.0",
    "elasticsearch==8.15.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pyhumps>=3.8.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "sqlmodel>=0.0.24",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/talent_platform"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pysocks>=1.7.1",
    "ruff>=0.12.0",
]
